Finished cleaner.
The Cleaner class handles deletion of outdated backup files according to configuration by days_to_keep, weeks_to_keep and months_to_keep. It is available under the easybackups_clean command. By default it prints files marked for deletion to stdout. If it should actually delete the files the command features a --dryrun option which needs to be set to False in order to remove the files. Resolves #7.